handling errors

+++ src/main/java/com/example/ticketingsystem/repository/TicketRepo.java
@@ -3,5 +3,5 @@ package com.example.ticketingsystem.repository;
 import com.example.ticketingsystem.models.Ticket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TicketRepo extends JpaRepository<Ticket, Integer> {
+public interface TicketRepo extends JpaRepository<Ticket, Long> {
 }
